Initial commit
Server (proxy.py) uses asyncio to create a proxy the Evolution Connex
thermostat can connect to through its proxy settings. The proxy intercepts
requests and responses and has the ability to send the data to another
part of the system for processing.

The server also uses aiohttp to create a WebSocket a client can connect
to to control what happens to all requests and responses. Commands and
results are sent as text. Payloads are sent as bytes.

A example command-line client (integration.py) uses aiohttp to connect to
the server's WebSocket and analyzes the traffic, converting data to JSON
and displaying on the screen.