Save a full conversation example
This was captured from a full conversation between the client (the
thermostat) and the server. "Full conversation" is the back and forth
that happens when the client reconnects to the internet after being
offline for some time. At that time, it will update the cloud with
everything it knows. All that chatter was captured here.