Make HttpServer callback optional
Previously ProxyServer was instantiating a dummy, pass-through callback
that would just forward requests to cover for the case where a callback
is not provided. Now, HttpServer just checks for the presence of a
callback and it will just forward the request/response if one is not
available.

Also, fixed a few mypy typing issues and in the process simplified the
callback interface to only return ProxyServerAction (instead of the
previous Union).