Allow for request suppression and custom responses
Callbacks can now suppress requests from being forwarded to the server.
In such cases, the callback is obligated to provide a response to the
client by itself, to close the loop. For that purpose, HttpResponse was
updated to allow it to be modified.

To avoid in the logic, the action taken by the request handling (Forward
or Suppress) is now sent as a parameter to the response callback as well.

Also in this change:

- Bumped version to 0.2.0
- Updated proxy callbacks signature so that responses don't return
  anything and the request action is passed in to inform response handling
- A few more mypy fixes
- Added a few more test tests